+++ service/controller/resource/monitoring/remotewriteingress/v1beta1/resource_test.go
@@ -1,0 +1,36 @@
+package v1beta1
+
+import (
+	"flag"
+	"path/filepath"
+	"testing"
+
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/unittest"
+)
+
+var update = flag.Bool("update", false, "update the ouput file")
+
+func TestIngress(t *testing.T) {
+	outputDir, err := filepath.Abs("./test")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	c := unittest.Config{
+		OutputDir: outputDir,
+		T:         t,
+		TestFunc: func(v interface{}) (interface{}, error) {
+			return toIngress(v, Config{BaseDomain: "https://prometheus"})
+		},
+		Update: *update,
+	}
+	runner, err := unittest.NewRunner(c)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	err = runner.Run()
+	if err != nil {
+		t.Fatal(err)
+	}
+}