+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Enable remote write receiver.
+- Add prometheus remote write ingress boilerplate.
 
 ## [4.5.1] - 2022-08-24